refs #4813 fixing unix warning

+++ Code/Mantid/Framework/Kernel/src/V3D.cpp
@@ -832,7 +832,7 @@ void V3D::loadNexus(::NeXus::File * file, const std::string & name)
   *
   *  As crystallographical coordinate sytem is based on 3 integers, eps is used as accuracy to convert into integers
 */
-#define DINT(x)    std::fabs((x)-size_t((x)+0.5))
+#define DINT(x)    std::fabs((x)-double(size_t((x)+0.5)))
 double nearInt(double val,double eps,double mult)
 {
    if(val>0){
